refactor: move `get_nwjs` and `get_ffmpeg` into one file
