+++ src/get.js
@@ -1,4 +1,5 @@
-import { createWriteStream } from "node:fs";
+import { spawnSync } from "node:child_process";
+import { createWriteStream, existsSync } from "node:fs";
 import { mkdir, readdir, rm, rmdir } from "node:fs/promises";
 import { get as getRequest } from "node:https";
 import { resolve } from "node:path";
@@ -9,7 +10,7 @@ import compressing from "compressing";
 
 import { log } from "./log.js";
 import { PLATFORM_KV, ARCH_KV } from "./util.js";
-import child_process from "child_process";
+import { replaceFfmpeg } from "./util/ffmpeg.js";
 
 /**
  * _Note: This an internal function which is not called directly. Please see example usage below._
@@ -46,6 +47,52 @@ import child_process from "child_process";
  *  downloadUrl: "https://cnpmjs.org/mirrors/nwjs/",
  * });
  *
+ * @example
+ * // FFMPEG (proprietary codecs)
+ * // Please read the license's constraints: https://nwjs.readthedocs.io/en/latest/For%20Developers/Enable%20Proprietary%20Codecs/#get-ffmpeg-binaries-from-the-community
+ * nwbuild({
+ *   mode: "get",
+ *   ffmpeg: true,
+ * });
+ *
+ * @param  {object}                   options              Get mode options
+ * @param  {string}                   options.version      NW.js runtime version. Defaults to "latest".
+ * @param  {"normal" | "sdk"}         options.flavor       NW.js build flavor. Defaults to "normal".
+ * @param  {"linux" | "osx" | "win"}  options.platform     Target platform. Defaults to host platform.
+ * @param  {"ia32" | "x64" | "arm64"} options.arch         Target architecture. Defaults to host architecture.
+ * @param  {string}                   options.downloadUrl  File server to download from. Defaults to "https://dl.nwjs.io". Set "https://npm.taobao.org/mirrors/nwjs" for China mirror or "https://cnpmjs.org/mirrors/nwjs/" for Singapore mirror.
+ * @param  {string}                   options.cacheDir     Cache directory path. Defaults to "./cache"
+ * @param  {boolean}                  options.cache        If false, remove cache before download. Defaults to true.
+ * @param  {boolean}                  options.ffmpeg       If true, ffmpeg is not downloaded. Defaults to false.
+ * @return {Promise<void>}
+ */
+export async function get({
+  version = "latest",
+  flavor = "normal",
+  platform = PLATFORM_KV[PLATFORM],
+  arch = ARCH_KV[ARCH],
+  downloadUrl = "https://dl.nwjs.io",
+  cacheDir = "./cache",
+  cache = true,
+  ffmpeg = false,
+}) {
+  await get_nwjs(version, flavor, platform, arch, downloadUrl, cacheDir, cache);
+  if (ffmpeg === true) {
+    await get_ffmpeg({
+      version,
+      flavor,
+      platform,
+      arch,
+      downloadUrl,
+      cacheDir,
+      cache,
+    });
+  }
+}
+
+/**
+ * Get NW.js binaries
+ *
  * @param  {object}                   options              Get mode options
  * @param  {string}                   options.version      NW.js runtime version. Defaults to "latest".
  * @param  {"normal" | "sdk"}         options.flavor       NW.js build flavor. Defaults to "normal".
@@ -56,7 +103,7 @@ import child_process from "child_process";
  * @param  {boolean}                  options.cache        If false, remove cache before download. Defaults to true.
  * @return {Promise<void>}
  */
-export async function get_nwjs({
+async function get_nwjs({
   version = "latest",
   flavor = "normal",
   platform = PLATFORM_KV[PLATFORM],
@@ -144,7 +191,7 @@ export async function get_nwjs({
               //TODO: compressing package does not restore symlinks on some macOS (eg: circleCI)
               const exec = function (cmd) {
                 log.debug(cmd);
-                const result = child_process.spawnSync(cmd, {
+                const result = spawnSync(cmd, {
                   shell: true,
                   stdio: "inherit",
                 });
@@ -181,4 +228,119 @@ export async function get_nwjs({
       log.debug(`Binary zip removed`);
     });
   }
+}
+
+/**
+ * Get FFmpeg binary.
+ *
+ * @param  {object}                   options           Get mode options
+ * @param  {string}                   options.version   NW.js runtime version. Defaults to "latest".
+ * @param  {"normal" | "sdk"}         options.flavor    NW.js build flavor. Defaults to "normal".
+ * @param  {"linux" | "osx" | "win"}  options.platform  Target platform. Defaults to host platform.
+ * @param  {"ia32" | "x64" | "arm64"} options.arch      Target architecture. Defaults to host architecture.
+ * @param  {string}                   options.cacheDir  Cache directory path. Defaults to "./cache"
+ * @param  {boolean}                  options.cache     If false, remove cache before download. Defaults to true.
+ * @return {Promise<void>}
+ */
+async function get_ffmpeg({
+  version = "latest",
+  flavor = "normal",
+  platform = PLATFORM_KV[PLATFORM],
+  arch = ARCH_KV[ARCH],
+  cacheDir = "./cache",
+  cache = true,
+}) {
+  log.debug(`Start getting binaries`);
+  const nwDir = resolve(
+    cacheDir,
+    `nwjs${flavor === "sdk" ? "-sdk" : ""}-v${version}-${platform}-${arch}`,
+  );
+  const bar = new progress.SingleBar({}, progress.Presets.rect);
+
+  // If options.ffmpeg is true, then download ffmpeg.
+  const downloadUrl =
+    "https://github.com/nwjs-ffmpeg-prebuilt/nwjs-ffmpeg-prebuilt/releases/download";
+  let url = `${downloadUrl}/${version}/${version}-${platform}-${arch}.zip`;
+  const out = resolve(cacheDir, `ffmpeg-v${version}-${platform}-${arch}.zip`);
+
+  // If options.cache is false, remove cache.
+  if (cache === false) {
+    log.debug(`Removing existing binaries`);
+    await rm(out, {
+      recursive: true,
+      force: true,
+    });
+  }
+
+  const unzipFFMPeg = async () => {
+    if (platform === "linux") {
+      await compressing.tgz.uncompress(out, nwDir);
+    } else {
+      await compressing.zip.uncompress(out, nwDir);
+    }
+    let ffmpegFile;
+    if (platform === "linux") {
+      ffmpegFile = "libffmpeg.so";
+    } else if (platform === "win") {
+      ffmpegFile = "ffmpeg.dll";
+    } else if (platform === "osx") {
+      ffmpegFile = "libffmpeg.dylib";
+    }
+    await replaceFfmpeg(platform, nwDir, ffmpegFile);
+
+    if (cache === false) {
+      log.debug(`Removing FFMPEG zip cache`);
+      await rm(out, {
+        recursive: true,
+        force: true,
+      });
+      log.debug(`FFMPEG zip cache removed`);
+    }
+  };
+  // Check if cache exists.
+  if (existsSync(out)) {
+    log.debug(`Found existing FFMPEG cache`);
+    await unzipFFMPeg();
+    return;
+  }
+
+  log.debug(`Downloading FFMPEG`);
+  const stream = createWriteStream(out);
+  const request = new Promise((resolve, reject) => {
+    getRequest(url, (response) => {
+      log.debug(`Response from ${url}`);
+      // For GitHub releases and mirrors, we need to follow the redirect.
+      url = response.headers.location;
+
+      getRequest(url, (response) => {
+        log.debug(`Response from ${url}`);
+        let chunks = 0;
+        bar.start(Number(response.headers["content-length"]), 0);
+        response.on("data", async (chunk) => {
+          chunks += chunk.length;
+          bar.increment();
+          bar.update(chunks);
+        });
+
+        response.on("error", (error) => {
+          reject(error);
+        });
+
+        response.on("end", () => {
+          log.debug(`FFMPEG fully downloaded`);
+          bar.stop();
+          resolve();
+        });
+
+        response.pipe(stream);
+      });
+
+      response.on("error", (error) => {
+        reject(error);
+      });
+    });
+  });
+
+  // Remove compressed file after download and decompress.
+  return request.then(unzipFFMPeg);
 }
